Moved worker client&server inside worker
