+++ pkg/worker/workerserver/artifact.go
